[PATCH] dvb: mt352 frontend driver update
This patch updates the mt352 driver to make it work with the Pinnacle 300i
card.  Some values are calculated at runtime now instead of having hard-coded
defaults because the defaults don't fit for the Pinnacle, and some more small
tweaks + fixes.

This was also discussed + accepted on the dvb list, going to submit directly
because the saa7134 driver update depends on this.

Signed-off-by: Gerd Knorr <kraxel@bytesex.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>