[PATCH] arch/sparc/kernel/ptrace: pointless assignment and shadowed var
A few tiny cleanups for arch/sparc/kernel/ptrace.c::do_ptrace post the big
verify_area/access_ok cleanup.

'ret' shadows a variable of the same name in the enclosing scope, rename it.

The assignment of -EFAULT to 'i' exactely mirrors what the old verify_area
code did, but that was only to use 'i' to check the return value of
verify_area.  Now that we check access_ok directly and 'i' is initialized in
the for loop a few lines below anyway, the asignment of -EFAULT to i is bogus,
just pass pass EFAULT directly as the second arg to pt_error_return. 

Also a few tiny whitespace cleanups - 'if ()' vs 'if()'.

Signed-off-by: Jesper Juhl <jesper.juhl@gmail.com>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>