[PATCH] ftape/fdc-io: insert set_current_state() before schedule_timeout()
Inserts a set_current_state(TASK_INTERRUPTIBLE) before the schedule_timeout()
call.  Without this change, after the first iteration of the loop,
schedule_timeout() will not only return immediately, but the loop will break,
as the conditional will no longer be satisfied.  In fact, this conditional
makes little sense given the workings of schedule_timeout.  The timeout
variable is ignored, as well, and I'm fairly certain that it should be
included in the loop conditional.  That way, if the timeout expires before a
signal hits, -ETIME will be returned by fdc_interrupt_wait() instead of
-EINTR.

Signed-off-by: Nishanth Aravamudan <nacc@us.ibm.com>
Signed-off-by: Domen Puncer <domen@coderock.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>