[PATCH] USB Storage: change how unusual_devs.h flags are defined
This patch started life as as473 from Alan Stern, and has been rediffed
against the current tip.

This patch changes the way the flags bits are defined, so that the
proc_info routine can automatically list all the flags that are set.  As a
side effect it will also include the NEED_OVERRIDE flag; that shouldn't
really matter.  (As another side effect, the definitions grew too long to
fit on a single 80-character line so I had to wrap the comments.)

This patch also takes the opportunity to remove a couple of obsolete flags
and to renumber the bit values.  The patch also removes an unusual_devs
entry whose only purpose was to set one of those obsolete flags; apparently
it got left behind when all the other uses of those flags were removed.

Signed-off-by: Alan Stern <stern@rowland.harvard.edu>
Signed-off-by: Matthew Dharm <mdharm-usb@one-eyed-alien.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>