[IPSEC]: Get ttl from child instead of path
Now that dst_pmtu is almost gone let's do the same to dst_path_metric.
I've only found one legitimate use of it and that's the one that was
recently added in dst_allfrag.

This patch makes xfrm4_encap/xfrm6_encap use dst->child instead of
dst->path so that we choose the correct route to get the hoplimit
from when nested tunnels are present.

For simple tunnels dst->child == dst->path so there is no change.

Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: David S. Miller <davem@davemloft.net>