[PATCH] ppc64: make cpu hotplug play well with maxcpus and smt-enabled
This patch allows you to boot a pSeries system with maxcpus=x or
smt-enabled=off (or both) and bring up the offline cpus later from userspace,
assuming the kernel was built with CONFIG_HOTPLUG_CPU=y.

- Record cpus which were started from OF in a cpu map and use that instead
  of system_state to decide how to start a cpu in smp_startup_cpu.

- Change the smp bootup logic slightly so that the path for bringing up
  secondary threads is exactly the same as hotplugging a cpu later from
  userspace.

- Add a new function to smp_ops - cpu_bootable.  This is implemented only by
  pSeries to filter out secondary threads during boot with smt-enabled=off.
  Another way this could be done is to change the kick_cpu member to return
  int and we can check for this case in smp_pSeries_kick_cpu.

- Remove the games we play with cpu_present_map and the
  hard_smp_processor_id to handle smt-enabled=off, since they're now
  unnecessary.

- Remove find_physical_cpu_to_start; assigning threads to logical slots
  should be done at bootup and at DLPAR time, not during a cpu online
  operation.

One caveat: you need up-to-date firmware on Power5 for the maxcpus option to
work on systems with more than one processor.  Otherwise interrupts get
misrouted, typically resulting in hangs or "unable to find root filesystem"
problems.

Tested on Power5 with and without CONFIG_HOTPLUG_CPU and with various
combinations of the maxcpus= and smt-enabled= parameters.

Signed-off-by: Nathan Lynch <ntl@pobox.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>