[PATCH] fix breakage in the SCSI generic tag code
When I redid the target allocation routines, I inadvertently broke the
generic tag code.  The problem is that the device is added to the host
list after slave_configure in the new code.  This is too late since if
any command disconnects during the initial inquiry phase, we now cannot
find the device for the reconnect.  The fix is to add the device to the
lists earlier.

Signed-off-by: James Bottomley <James.Bottomley@SteelEye.com>