ncr53c8xx: Fix small problem with initial negotiation
The driver produces messages list this:

ncr53c720-1: ID 7, Fast-10, Parity Checking
scsi1 : ncr53c8xx-3.4.3g
 target1:0:1: target did not report SYNC.
  Vendor: TOSHIBA   Model: CD-ROM XM-5401TA  Rev: 3605
  Type:   CD-ROM                             ANSI SCSI revision: 02
 target1:0:1: Beginning Domain Validation
 target1:0:1: asynchronous.
 target1:0:1: Domain Validation skipping write tests
 target1:0:1: FAST-5 SCSI 4.2 MB/s ST (236 ns, offset 8)
 target1:0:1: Ending Domain Validation

Although everything is fine.  The initial "target did not report SYNC"
is because the driver is trying to negotiate over the initial inquiry
(so, naturally we don't know if it supports sync yet).  The solution is
to rip out the last remnants of the spontaneous attempts at negotiation
to force it to wait for the DV negotiation in slave_configure.

Signed-off-by: James Bottomley <James.Bottomley@SteelEye.com>