[ARM PATCH] 2573/1: simplify align[bw]() in ixp2000's io.h and update comments
Patch from Lennert Buytenhek

We currently have a workaround for the fact that early IXP2400s
assert byte lanes for PCI I/O transactions the other way round.
Newer silicon has a bit that can be set to 1 to get the 'proper'
behavior, but since that is not available on older silicon we
choose to keep working around this issue by hand.
The workaround can be done in an easier way than is currently the
case, though -- instead of substracting the lower two bits of the
I/O address from 3 and putting the result back into the lower two
bits, we can just toggle the two lowest bits using a simple XOR
with 3.
While we are at it, the outb() macro does not seem to parenthesize
its first argument, while outw() and outl() do, so fix this up.

Signed-off-by: Lennert Buytenhek
Signed-off-by: Deepak Saxena
Signed-off-by: Russell King