[TG3]: Flush status block in tg3_interrupt()
Add register read of PCI state register in tg3_interrupt() if status block's
updated bit is not set. This will flush the status block and confirm whether
the interrupt is ours or not. PCI ordering rules allow the interrupt to
arrive at the CPU ahead of the status block that may be posted at the
chipset.

Signed-off-by: Michael Chan <mchan@broadcom.com>
ACKed-by: Jeff Garzik <jgarzik@pobox.com>
Signed-off-by: David S. Miller <davem@davemloft.net>