[CRYPTO]: Kill obsolete iv check in cbc_process()
Here's some more optimisations plus a bug fix for a pathological case
where in_place might not be set correctly which can't happen with any
of the current users.  Here is the first one:

We have long since stopped using a null cit_iv as a means of doing null
encryption.  In fact it doesn't work here anyway since we need to copy
src into dst to achieve null encryption.

No user of cbc_encrypt_iv/cbc_decrypt_iv does this either so let's just
get rid of this check which is sitting in the fast path.

Signed-off-by: David S. Miller <davem@davemloft.net>