[PATCH] USB: usbnet minor bugfixes
Two bugfixes to usbnet.  The important one is that it's OK when the
minidriver doesn't want to use a status endpoint.  That shouldn't
be treated as an error ... errors prevent probe() from succeeding!

The minor fix is that software driven interface shutdown should neither
be accounted as an error, nor generate a diagnostic.

Signed-off-by: David Brownell <dbrownell@users.sourceforge.net>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>