[PATCH] USB Storage: Remove dup in unusual_devs
Matthew Dharm pointed out that your BK tree somehow got 2 versions of an
unusual_devs entry in it. One was from a patch in January and the other
was from an automerge.

I dug around in bk - which I find to be the most counter-intuitive
program in the world - and generated a patch against your latest BK
tree, and I've attached it. It just removes the dup.

Signed-off-by: Phil Dibowitz <phil@ipom.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@suse.de>