[PATCH] ppc64: introduce pSeries_reconfig.[ch]
Move as much pSeries-specific DLPAR/hotplug code as possible into its own
file, which is built only when pSeries support is enabled in the config.  This
new file is intended to contain support code for the "Dynamic Reconfiguration"
option in the RISC Platform Architecture, which encompasses both PCI hotplug
and dynamic logical partitioning (DLPAR).

This patch mostly just moves code around, but the device node addition and
removal API is slightly modified.  In this way, of_add_node and of_remove_node
are now responsible only for safely updating the device tree and global list,
without all the other stuff like proc entries etc.  of_add_node and
of_remove_node have been renamed to of_attach_node and of_detach_node,
respectively.

This also adds the definitions and api for a notifier chain which is meant to
be used by code that must act upon device node addition or removal.  Patches
to migrate code to the notifier api follow in this series.

Signed-off-by: Nathan Lynch <ntl@pobox.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>