[ARM PATCH] 2576/1: Fix LDRD and LDRSB (Thumb) abort handling
Patch from Deepak Saxena

The ARM LDRD (v5+) and Thumb LDRSB instructions use bit 20/11 (ARM/Thumb)
differently than every other instruction, so it is set to 0 (write) even
though the instructions are read instructions. This means that during an
abort the instructions will be treated as a write and the handler will
raise a signal from unwriteable locations if they fault. We have to
specifically check for these instructions from the abort handlers to
treat them properly. EABI toolchains emit LDRD instructions and even
those not using EABI might have handcoded ASM that uses these instructions.
This patch creates an abort-macro.S file that is included by the
abort-ev* files that need the special cases and changes those
handlers to use the macros.  ARMv6 does not need the special case
handling as the HW takes care of setting the FSR bit appropriately.

Signed-off-by: Deepak Saxena
Signed-off-by: Russell King