[PATCH] ppc64: Fix ethernet PHY reset on iMac G5
On iMac G5, when netbooting (or causing any other ethernet activity from
within the Open Firmware environment), the PHY is put into a low power
state before booting the OS.  The result is that Linux doesn't see it and
networking doesn't work.

This patch adds the ethernet PHY reset platform hook to pmac_feature.c on
ppc64 (it already is commonly used on ppc32 as lots of macs have this same
problem, so the hook definition is already there and sungem is already
calling it).

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@osdl.org>
Signed-off-by: Linus Torvalds <torvalds@osdl.org>